GA now introduces random DNA into population
